receiver/requirements.txt
spidev
pycrypto
grpcio
protobuf
